+++ ShootServWeb/Models/ShootingRange/ShootingRangeModelParams.cs
@@ -2,8 +2,6 @@
 using System.ComponentModel;
 using System.ComponentModel.DataAnnotations;
 using System.Web.Mvc;
-using BL;
-using ShootingCompetitionsRequests.Models;
 
 namespace ShootServ.Models.ShootingRange
 {
@@ -78,9 +76,8 @@ namespace ShootServ.Models.ShootingRange
 
         public ShootingRangeModelParams()
         { 
-            Countries = StandartClassifierModelLogic.GetCountryList();
-            Regions = StandartClassifierModelLogic.GetRegionsByCountry(1); /// TODO : HardCode
+            Countries = new List<SelectListItem>();
+            Regions = new List<SelectListItem>();
         }
-
     }
 }
