+++ ShootServWeb/Scripts/ShootServ/shootingRange.js
@@ -1,46 +1,11 @@
 $(document).ready(function ()
 {
     var actor = new shootingRangePageActor();
-
-    actor.changeRegion();
-    $(document).on("change", "#region-choise", function ()
-    {
-        actor.changeRegion();
-    });
-
+    
     $(document).on("change", "#CountryId", function () {
         actor.changeCountry();
     });
-
-    // Клик на ссылке удалить тир
-    $(document).on("click", ".delRange", function () {
-
-        if ( isAuthorize) {
-            var a = $(this);
-            var tr = a.closest("tr");
-            var id = a.attr("idShootRange");
-
-            $.ajax({
-                url: linksShootingRange.Delete,
-                dataType: "json",
-                data: {idShootingRange: id},
-                async: false,
-                success: function (data) {
-                    if (data.IsOk) {
-                        showInfo("Тир удален");
-                        tr.remove();
-                    } else showError(data.Message);
-                },
-                error: function (data) {
-                    showError("Ошибка ajax");
-                }
-            });
-        }
-        else {
-            redirectToLoginPage(linksShootingRange.Index);
-        }
-    });
-
+    
     $(document).on("click", "#addBt", function ()
     {
         if( isAuthorize) {
@@ -53,13 +18,12 @@ $(document).ready(function ()
                     async: false,
                     success: function (data) {
                         if (data.IsOk) {
-                            var idRegion = $("#RegionId").val();
-                            actor.getListByRegion(idRegion);
+                            showInfo("Тир успешно добавлен");
+                            window.location = linksShootingRange.List;
                         } else {
                             if (data.Message) {
                                 showError(data.Message); // сообщение об ошибке как -то показать на странице 
                             }
-                            console.dir(data.ValidateMessages);
                             // чтобы было совсем хорошо
                             window.common.lightValidationMessagesOnForm($("form"), data.ValidationMessages);
                         }
@@ -81,13 +45,6 @@ $(document).ready(function ()
 var shootingRangePageActor = function () {
     
     // Нужно вызвать эту функцию при изменении региона
-    this.changeRegion = function()
-    {
-        var idRegion = $("#region-choise").val();
-        this.getListByRegion(idRegion);
-    };
-    
-    // Нужно вызвать эту функцию при изменении региона
     this.changeCountry = function() {
         var idCountry = $("#CountryId").val();
         if (idCountry) {
@@ -95,7 +52,7 @@ var shootingRangePageActor = function () {
             var idRegion = $("#RegionId");
             renderJsonArrayToSelect(idRegion, "Id", "Name", regions);
         }
-        };
+    };
 
     this.construct = function(){
     };
@@ -122,26 +79,4 @@ var shootingRangePageActor = function () {
     
         return true;
     };
-    
-    // Получить список тиров по региону
-    this.getListByRegion = function(idRegion)
-    {
-        if (idRegion) {
-            $.ajax({
-                url: linksShootingRange.GetListByRegion,
-                dataType: "html",
-                data: { idRegion: idRegion },
-                async: false,
-                success: function (data) {
-                    /// TODO: в data json, надо построить в listShootingRanges таблицу
-                    $("#listShootingRanges").html(data); 
-                },
-                error: function ()
-                {
-                    showError("Ошибка ajax");
-                }
-            });
-        }
-    };
-
 };
