+++ ShootServWeb/Views/Shared/_Layout.cshtml
@@ -54,7 +54,7 @@
                         <div class="dropdown-menu" aria-labelledby="dropdownMenuLink">
                             <a class="p-2 text-muted dropdown-item" href="@Url.Action("Index", "ShootingRange")">Добавление тиров</a>
                             <div class="dropdown-divider"></div>
-                            <a class="p-2 text-muted dropdown-item" href="@Url.Action("Index", "ShootingRange")">Просмотр тиров</a>
+                            <a class="p-2 text-muted dropdown-item" href="@Url.Action("List", "ShootingRange")">Просмотр тиров</a>
                         </div>
                     </div>
                       
@@ -108,6 +108,7 @@
         <script type="text/javascript">
             var linksShootingRange = {
                 Index: "@Url.Action("Index", "ShootingRange")",
+                List : "@Url.Action("List", "ShootingRange")", 
                 Add: "@Url.Action("Add", "ShootingRange")",
                 Delete: "@Url.Action("Delete", "ShootingRange")",
                 GetListByRegion: "@Url.Action("GetListByRegion", "ShootingRange")"
