+++ ShootServWeb/Views/ShootingRange/Index.cshtml
@@ -12,20 +12,43 @@
 <div id="divShootingRange">
     <div id="divDiv" class="instr">
         <div id="divAddShootingRange" class="addListStyle"></div>
-            @Html.Partial("Add", Model)
-    </div>
-</div>
-
- @Html.Label("region-choise", "Регион")
- @Html.DropDownList("region-choise", Model.Regions)
-
- <div id="list" style="margin-top: 15px; width: 100%; height: 80%;">
-    <div id="divDiv1" class="instr" style="width: 1000px; padding: 0;">
-        <div id="divShowShootingRange" class="addListStyle" style="width: 100%; padding-bottom: 20px; padding-top: 20px;"></div>
-        <div id="divShowShootingR">
-            <div id="listShootingRanges">
+        @using (Html.BeginForm("Add", "ShootingRange", FormMethod.Post, new { id = "addform" }))
+        {
+            <div class="mt-4">
+                @Html.LabelFor(x => x.Name)
+                @Html.TextBoxFor(x => x.Name, new { @class = "form-control" })
             </div>
-        </div>
+            <div class="row mt-4">
+                <div class="col-lg-6 col-xs-12">
+                    @Html.LabelFor(x => x.CountryId)
+                    @Html.DropDownListFor(x => x.CountryId, Model.Countries, new { @class = "form-control" })
+                </div>
+                <div class="col-lg-6 col-xs-12">
+                    @Html.LabelFor(x => x.RegionId)
+                    <div id="tdRegionId">@Html.DropDownListFor(x => x.RegionId, Model.Regions, new { @class = "form-control" })</div>
+                </div>
+            </div>
+            <div class="row mt-4">
+                <div class="col-lg-6 col-xs-12">
+                    @Html.LabelFor(x => x.Town)
+                    @Html.TextBoxFor(x => x.Town, new { @class = "form-control" })
+                </div>
+                <div class="col-lg-6 col-xs-12">
+                    @Html.LabelFor(x => x.Phone)
+                    @Html.TextBoxFor(x => x.Phone, new { @class = "form-control" })
+                </div>
+            </div>
+            <div class="mt-4">
+                @Html.LabelFor(x => x.Address)
+                @Html.TextAreaFor(x => x.Address, new { @class = "form-control" })
+            </div>
+            <div class="mt-4">
+                @Html.LabelFor(x => x.Info)
+                @Html.TextAreaFor(x => x.Info, new { @class = "form-control" })
+
+            </div>
+            <input id="addBt" class="btn btn-lg btn-primary mx-auto mt-4" type="button" value="Добавить" />
+        }
     </div>
 </div>
