+++ ShootServWeb/Views/ShootingRange/List.cshtml
@@ -1,0 +1,14 @@
+<script src="~/Scripts/ShootServ/shootingRangeList.js"></script>
+
+@Html.Label("region-choise", "Region")
+@Html.DropDownList("region-choise", (List<SelectListItem>)ViewBag.Regions)
+
+<div id="list" style="margin-top: 15px; width: 100%; height: 80%;">
+    <div id="divDiv1" class="instr" style="width: 1000px; padding: 0;">
+        <div id="divShowShootingRange" class="addListStyle" style="width: 100%; padding-bottom: 20px; padding-top: 20px;"></div>
+        <div id="divShowShootingR">
+            <div id="listShootingRanges">
+            </div>
+        </div>
+    </div>
+</div>
